+++ core/specfem/medium_physics.hpp
@@ -22,7 +22,7 @@ namespace specfem::medium {}
 #include "medium_physics/compute_coupling.hpp"
 #include "medium_physics/compute_damping_force.hpp"
 #include "medium_physics/compute_frechet_derivatives.hpp"
-#include "medium_physics/compute_mass_matrix.hpp"
-#include "medium_physics/compute_source.hpp"
+#include "medium_physics/compute_source_contribution.hpp"
 #include "medium_physics/compute_stress.hpp"
 #include "medium_physics/compute_wavefield.hpp"
+#include "medium_physics/mass_matrix_component.hpp"
