+++ core/specfem/medium_physics/compute_source_contribution.hpp
