+++ core/specfem/medium_physics/mass_matrix_component.hpp
