+++ docs/sections/api/medium/cosserat_coupled_stress/index.rst
@@ -1,9 +1,9 @@
 .. _api_medium_cosserat_couple_stress:
 
-``specfem::medium::compute_couple_stress``
-===========================================
+``specfem::medium::compute_cosserat_couple_stress``
+===================================================
 
-.. doxygenfunction:: specfem::medium::compute_couple_stress
+.. doxygenfunction:: specfem::medium::compute_cosserat_couple_stress
 
 Implementation Details
 ----------------------
